+++ contrib/hourglass/src/java/datafu/hourglass/jobs/MaxInputDataExceededException.java
@@ -1,0 +1,13 @@
+package datafu.hourglass.jobs;
+
+public class MaxInputDataExceededException extends Throwable
+{
+  public MaxInputDataExceededException()
+  {    
+  }
+  
+  public MaxInputDataExceededException(String message)
+  {
+    super(message);
+  }
+}